Merge unit tests of ImageData and Image classes
ImageData and Image classes are merged into a single class in a prior
commit. Similarly, this patch moves the tests from ImageData_test
class to Image_test class.

Number of total tests are reduced from 60 (#ImageClass_test + #Image_test)
to 53. This is due to duplicate tests that are removed. Below is a list
of tests that are removed, and the tests they're replaced with.

======================================================
|      Removed            |      Replaced with       |
======================================================
|ImageData_test           | Image_test               |
------------------------------------------------------
|GetImageIDandImageFormat | GetMatFromTDB            |
|GetImageDimensions       | GetImageDimensions       |
|GetEncodedBuffer         | EncodedImage             |
|CreateUnique             | CreateNamePNG            |
======================================================

The remaining 3 tests had the same name in both files, so they're not
shown in the list.

Signed-off-by: mahircg <mahircan.guel@intel.com>