Merge VideoData and Video classes
VideoData class was used as a wrapper around Video class which was
maintained in a seperate repository, but not any longer. Since Video
class is part of the VDMS codebase, abstraction layer provided by
VideoData is redundant.

This patch combines VideoData and Video classes into the Video class to
remove the interface layer in between.

Signed-off-by: mahircg <mahircan.guel@intel.com>