Add key frame decoding method to Video class
This patch introduces 'get_decoded_key_frames' method to Video class.
The method receives a list of key frame indexes and indexes of frames to
return, and invokes partial decoding of the video with those parameters.

This method is meant to be used when the video node has key-frame
information associated with it, that can be exploited for partial
decoding rather than the full decoding process.

Signed-off-by: mahircg <mahircan.guel@intel.com>