Add KeyFrameOp and KeyFrameDecoder classes
This patch introduces 'KeyFrameDecoder' class, that can be used for
partial video decoding. The class is initialized with a list of
key-frame index/offset pairs, to be used for performing the partial
video decoding.

Moreover, both KeyFrameParser and KeyFrameDecoder classes employ a
common set of structures for initialization. Thus, a base class named
'KeyFrameOp' is added, to be inherited by both parsing and decoding
classes.