196 objects not cleaned up on transaction failure (#306)
* Include S3 transaction rollback test

This test counts the number of objects stored in S3, then executes a query
that would cause a storage leak and finally makes sure the number of objects
stored in S3 is still the same.

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>

* Implement S3 transaction rollback for PMGD handler

In case of failure, the exception handler will make sure all added objects
within the current query context are removed from S3.

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>

* Implement S3 transaction rollback for Neo4j handler

In case of failure, all added objects within the current query context are removed from S3.

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>

* Automated coverage update

---------

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>