Update QueryNode and QueryEdge protobuf construct handler read mode to avoid internal PMGD validation check exception
A read-write validation check is performed internally by PMGD. When running a query node/edge
transaction with a readOnly flag a validation check will trigger an exception.
The exception message will be propagated to the query response.

This change allows the validation to pass while technically the transaction is still a readOnly operation.
The validation occurs in the PMGD codebase, updating it is not reasonable at this time.

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>