182 erroneous read only on server start (#308)
* Update ImageTransactionRollback test and include new test FindImageEmptyDB

This test runs a query with 2 FindImage commands on a new/empty PMGD db instance
and validates the response contains the excepted message `"No entities found"`.

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>

* Update QueryNode and QueryEdge protobuf construct handler read mode to avoid internal PMGD validation check exception

A read-write validation check is performed internally by PMGD. When running a query node/edge
transaction with a readOnly flag a validation check will trigger an exception.
The exception message will be propagated to the query response.

This change allows the validation to pass while technically the transaction is still a readOnly operation.
The validation occurs in the PMGD codebase, updating it is not reasonable at this time.

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>

* Restore previous test changes

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>

---------

Signed-off-by: Steven Rojas <steven.rojas.cubero@intel.com>